update blog ,trait,category

+++ app/Traits/TraitCRUD.php
@@ -46,7 +46,12 @@ trait TraitCRUD
     public function store(Request $request)
     {
         $data = $request->all();
-      
+
+        if (!isset($data['slug'])) {
+            $data['slug'] = Str::slug($request['title']);
+        }
+
+
         foreach ($data as $key => $value) {
             if (Str::startsWith($key, 'is_')) {
                 $data[$key] = $request->input($key);
@@ -70,7 +75,7 @@ trait TraitCRUD
                 $query->with($this->relations);
             })
             ->findOrFail($id);
-        return view('admin.' . $this->model->getTable() . '.' . __FUNCTION__, compact('data'));
+        return view('admin.' . $this->model->getTable() . '.' . __FUNCTION__, compact('data', 'dataID'));
     }
     public function edit($id)
     {
@@ -98,13 +103,26 @@ trait TraitCRUD
             } elseif (Str::startsWith($key, 'img_')) {
                 $data[$key] = $dataID->$key;
                 if ($request->hasFile($key)) {
-                    Storage::delete($dataID->$key);
+                    if ($dataID->$key && Storage::exists($dataID->$key)) {
+                        Storage::delete($dataID->$key);
+                    }
+
                     $data[$key] = Storage::put($this->model->getTable(), $request->file($key));
                 }
             }
         }
 
         $this->model->findOrFail($id)->update($data);
+
+        if (array_key_exists('is_active', $data)) {
+            // Nếu trường is_active tồn tại và được cập nhật thành false
+            if (!$data['is_active']) {
+                $dataID->children()->update(['is_active' => false]);
+            } else {
+                $dataID->children()->update(['is_active' => true]);
+            }
+        }
+
         return redirect()->route($this->model->getTable() . '.index')->with('success', __('Cập nhật dữ liệu thành công'));
     }
     public function destroy($id)
@@ -122,9 +140,9 @@ trait TraitCRUD
 
         $dataID = $this->model->withTrashed()->findOrFail($id);
 
-        // if (Storage::exists($dataID->img_path)) {
-        //     Storage::delete($dataID->img_path);
-        // }
+        if (!empty($dataID->img_path) && Storage::exists($dataID->img_path)) {
+            Storage::delete($dataID->img_path);
+        }
         $dataID->forceDelete();
         return redirect()->back()->with('success', __('Xóa vĩnh viễn dữ liệu thành công'));
     }
